# PERF: IntegerIndex._shallow_copy (and by extension, indexing)